requirements.txt
boto3==1.37.38
numpy==2.3.4
pandas==2.3.3
matplotlib==3.10.7
geopandas==1.1.1
pillow==12.0.0
torch==2.8.0
torchvision==0.23.0
fastai==2.8.4
timm==1.0.20
scikit-learn==1.7.2
albucore==0.0.24
albumentations==2.0.8